+++ mongoose.h
@@ -483,7 +483,7 @@ int c_vsnprintf(char *buf, size_t buf_size, const char *format, va_list ap);
 #if !(_XOPEN_SOURCE >= 700 || _POSIX_C_SOURCE >= 200809L) &&    \
         !(__DARWIN_C_LEVEL >= 200809L) && !defined(RTOS_SDK) || \
     defined(_WIN32)
-int strnlen(const char *s, size_t maxlen);
+size_t strnlen(const char *s, size_t maxlen);
 #endif
 
 #ifdef __cplusplus
